initial commit, utils file
